requirements.txt
fastapi~=0.115.11
JSON-log-formatter~=1.1.1
psutil~=7.0.0
dependency-injector~=4.46.0
PyYAML~=6.0.2
pydantic~=2.10.6
pytest~=8.3.5